ext/openssl: make OpenSSL::SSL::SSLSocket non-reusable
Fix the segmentation fault due to use after free of SSL object.

Don't free the SSL object in SSLSocket#stop and leave it in RTypedData.

OpenSSL::SSL::SSLSocket#stop (and the wrapper method #sysclose) was
documented as "Stops the SSL connection and prepares it for another
connection" and this patch changes it: once the SSLSocket stopped, it
can no longer hold a new SSL connection.

The method used to free the SSL object immediately after SSL_shutdown(),
but it is problematic. Since some methods, such as SSLSocket#connect
release the GVL while waiting the peer, there is a good chance that the
SSL object is freed from another thread and this leads to a segmentation
fault.