test/openssl: don't test default session timeout
The default session timeout for TLSv1 is 7200 and shouldn't be 300. And
this should not be checked because the value is decided by just "the 24
hours mentioned in the TLSv1 spec is way too long for http, the cache
would over fill" (from OpenSSL's source comment).

Old OpenSSL (<= 1.0.2) set ssl_ctx->session_timeout on SSL_CTX_new(),
which we call always with SSLv23_method(), and it isn't updated with
SSL_set_ssl_method().