string.c: avoid integer overflow caused by addition
Overflow of signed integer leads to undefined behavior. For instance,
rb_str_resize() has the issue. It uses a condition to check if the new
string can be embedded in the RString or not:

  else if (len + termlen <= RSTRING_EMBED_LEN_MAX + 1) {

This is problematic because the type of len is 'long' and termlen is an
'int'. On platform with sizeof(long)==4, it's fairly possible that len
become a large number like LONG_MAX-1 and termlen is larger than 1 for
multi-byte encodings like UTF-32.

Test cases are not included in this commit, as testing them would
consume huge (> 4GB) memory. But some of them can be triggered from Ruby
code:

  ruby -e'"x"<<"Z"*((1<<31)-3)'
  ruby -e'("x"*((1<<31)-2)).force_encoding("UTF-32").b'
  ruby -e'("x"*((1<<31)-2)).force_encoding("UTF-32").freeze'

* string.c (str_replace_shared_without_enc, str_new_frozen,
  str_make_independent_expand, rb_str_resize,
  rb_str_change_terminator_length): As described above, avoid integer
  overflow that can occur on 'len+termlen'.
  [ruby-core:XXXXX] [Bug #XXXXX]

* string.c (str_buf_cat): Fix so that it does not create a negative
  length String. Also fix the condition for 'string sizes too big' error
  so that it considers the non-one termlen.

* string.c (rb_str_modify_expand): Check that the new length is in the
  range of long before resizing. Also refactor to use RESIZE_CAPA_TERM
  macro.